requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
idna==2.9
iso8601==0.1.12
paho-mqtt==1.5.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-toolbelt==0.9.1
sqlparse==0.3.1
urllib3==1.25.8
wiotp-sdk==0.11.0
